+++ service/controller/app/resource/chart/desired.go
@@ -22,6 +22,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/giantswarm/app-operator/v6/pkg/project"
+	"github.com/giantswarm/app-operator/v6/pkg/status"
 	"github.com/giantswarm/app-operator/v6/service/controller/app/controllercontext"
 	"github.com/giantswarm/app-operator/v6/service/internal/indexcache"
 )
@@ -84,6 +85,7 @@ func (r *Resource) GetDesiredState(ctx context.Context, obj interface{}) (interf
 			}
 		}
 		if err != nil {
+			setStatus(cc, err)
 			return nil, microerror.Mask(err)
 		}
 	} else if err != nil {
@@ -246,15 +248,16 @@ func (r *Resource) buildTarballURL(ctx context.Context, cc *controllercontext.Co
 		r.logger.Errorf(ctx, err, "failed to get index.yaml from %q", repositoryURL)
 	}
 	if index == nil {
-		return "", "", microerror.Maskf(notFoundError, "index %#v for %q is <nil>", index, repositoryURL)
+		return "", "", microerror.Maskf(catalogEmptyError, "index %#v for %q is <nil>", index, repositoryURL)
 	}
 	if len(index.Entries) == 0 {
-		return "", "", microerror.Maskf(notFoundError, "index %#v for %q has no entries", index, repositoryURL)
+		return "", "", microerror.Maskf(catalogEmptyError, "index %#v for %q has no entries", index, repositoryURL)
 	}
 
 	entries, ok := index.Entries[cr.Spec.Name]
 	if !ok {
-		return "", "", microerror.Maskf(notFoundError, "no entries for app %#q in index.yaml for %q", cr.Spec.Name, repositoryURL)
+		// App is not there
+		return "", "", microerror.Maskf(appNotFoundError, "no entries for app %#q in index.yaml for %q", cr.Spec.Name, repositoryURL)
 	}
 
 	// We first try with the full version set in .spec.version of the app CR.
@@ -272,7 +275,7 @@ func (r *Resource) buildTarballURL(ctx context.Context, cc *controllercontext.Co
 	}
 
 	if url == "" {
-		return "", "", microerror.Maskf(notFoundError, "found entry for app %#q but URL is not specified", cr.Spec.Name)
+		return "", "", microerror.Maskf(appVersionNotFoundError, "found entry for app %#q but URL is not specified", cr.Spec.Name)
 	}
 
 	if !isValidURL(url) {
@@ -433,14 +436,14 @@ func getEntryURL(entries []indexcache.Entry, app, version string) (string, error
 	for _, e := range entries {
 		if e.Version == version {
 			if len(e.Urls) == 0 {
-				return "", microerror.Maskf(notFoundError, "no URL in index.yaml for app %#q version %#q", app, version)
+				return "", microerror.Maskf(appVersionNotFoundError, "no URL in index.yaml for app %#q version %#q", app, version)
 			}
 
 			return e.Urls[0], nil
 		}
 	}
 
-	return "", microerror.Maskf(notFoundError, "no app %#q in index.yaml with given version %#q", app, version)
+	return "", microerror.Maskf(appVersionNotFoundError, "no app %#q in index.yaml with given version %#q", app, version)
 }
 
 func hasConfigMap(cr v1alpha1.App, catalog v1alpha1.Catalog) bool {
@@ -528,4 +531,15 @@ func isOCIRepositoryURL(repositoryURL string) bool {
 		return false
 	}
 	return u.Scheme == "oci"
+}
+
+func setStatus(cc *controllercontext.Context, err error) {
+	switch microerror.Cause(err) {
+	case appNotFoundError:
+		addStatusToContext(cc, err.Error(), status.AppNotFoundStatus)
+	case appVersionNotFoundError:
+		addStatusToContext(cc, err.Error(), status.AppVersionNotFoundStatus)
+	case catalogEmptyError:
+		addStatusToContext(cc, err.Error(), status.CatalogEmptyStatus)
+	}
 }