+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 ## [Unreleased]
 
+### Changed
+
+- Improved feedback when working with looking for an app in catalog.
+
 ## [6.6.3] - 2023-03-21
 
 ## [6.6.3] - 2023-03-21